+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.02.20
+
+### Added
+
+- Designed a proof of concept for Depth Chart component
+
+### Updated
+
+- Changed design to include addPlayer, removePlayer, numPosition, numPlayers, and iterable methods
+
 ## [Unreleased]
 
 ## [2024.12.30]